requirements.txt
streamlit
tensorflow
pillow
scikit-learn
numpy
PyDrive2